+++ CHANGES.md
@@ -3,22 +3,28 @@
 ## Sound engine
 
 1. Feat: Sound can play many sounds (polyphonic) now.
+2. Fix: Sequence container - it should play sound correctly.
+3. Refactor: Rewrite Sound object without Pizzicato.Sound dependency.
 
 # Changes in v0.8.1
 
 ## Database
+
 1. Feat: Add new sound and set.
 
 ## Sound engine
+
 2. Fix: fields of containers randomize on each play event in loop.
 3. Fix: setting of non audio param field in sound objects is working.
 4. Fix: increased transition between new random values from 0 to 250ms.
 
 ## Main UI
+
 5. Feat: Add cover on the left side of the sound name in transport bar.
 6. Fix: sound name displaying in transport bar.
 7. Feat: Change sizes of cards.
 8. Fix: ASMR button follows play button in vertical direction.
 
 # Changes in v0.8.0
+
 This is the first beta version of this app It contains Main UI, prototype of Editor UI, almost done Sound Engine and AI Composer. There is also one sound for testing.